+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/mock/Expectation.java
@@ -7,6 +7,8 @@ package de.gematik.test.tiger.mockserver.mock;
 import de.gematik.test.tiger.mockserver.mock.action.ExpectationCallback;
 import de.gematik.test.tiger.mockserver.mock.action.ExpectationForwardAndResponseCallback;
 import de.gematik.test.tiger.mockserver.model.*;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.regex.PatternSyntaxException;
@@ -16,6 +18,7 @@ import lombok.Setter;
 import lombok.experimental.Accessors;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.http.client.utils.URIBuilder;
 
 /*
  * @author jamesdbloom
@@ -36,65 +39,6 @@ public class Expectation extends ObjectWithJsonToString implements Comparable<Ex
   private final HttpRequest requestPattern;
   @Setter private HttpAction httpAction;
   private ExpectationCallback expectationCallback;
-
-  /**
-   * Specify the HttpRequest to match against as follows:
-   *
-   * <p>
-   *
-   * <pre>
-   *     when(
-   *         request()
-   *             .withMethod("GET")
-   *             .withPath("/some/path")
-   *     ).thenRespond(
-   *         response()
-   *             .withContentType(APPLICATION_JSON_UTF_8)
-   *             .withBody("{\"some\": \"body\"}")
-   *     );
-   * </pre>
-   *
-   * <p>
-   *
-   * @param httpRequest the HttpRequest to match against
-   * @return the Expectation
-   */
-  public static Expectation when(HttpRequest httpRequest) {
-    return new Expectation(httpRequest);
-  }
-
-  /**
-   * Specify the HttpRequest to match against with a match priority as follows:
-   *
-   * <p>
-   *
-   * <pre>
-   *     when(
-   *         request()
-   *             .withMethod("GET")
-   *             .withPath("/some/path"),
-   *         10
-   *     ).thenRespond(
-   *         response()
-   *             .withContentType(APPLICATION_JSON_UTF_8)
-   *             .withBody("{\"some\": \"body\"}")
-   *     );
-   * </pre>
-   *
-   * <p>
-   *
-   * @param httpRequest the HttpRequest to match against
-   * @param priority the priority with which this expectation is used to match requests compared to
-   *     other expectations (high first)
-   * @return the Expectation
-   */
-  public static Expectation when(HttpRequest httpRequest, int priority) {
-    return new Expectation(httpRequest, priority);
-  }
-
-  public Expectation(HttpRequest requestDefinition) {
-    this(requestDefinition, 0);
-  }
 
   public Expectation(HttpRequest requestDefinition, int priority) {
     // ensure created enforces insertion order by relying on system time, and a counter
@@ -172,6 +116,36 @@ public class Expectation extends ObjectWithJsonToString implements Comparable<Ex
 
   @Override
   public int compareTo(Expectation o) {
+    if (o == null) {
+      return 1;
+    }
+    if (priority == o.priority) {
+      if (requestPattern == null
+          || o.requestPattern == null
+          || requestPattern.getPath() == null
+          || o.requestPattern.getPath() == null) {
+        return 0;
+      }
+      final String thisPath = requestPattern.getPath();
+      final String otherPath = o.requestPattern.getPath();
+      if (uriTwoIsBelowUriOne(thisPath, otherPath)) {
+        return -1;
+      } else if (uriTwoIsBelowUriOne(otherPath, thisPath)) {
+        return 1;
+      } else {
+        return 0;
+      }
+    }
     return Integer.compare(o.priority, priority);
+  }
+
+  private static boolean uriTwoIsBelowUriOne(final String value1, final String value2) {
+    try {
+      final URI uri1 = new URI(value1);
+      final URI uri2WithUri1Scheme = new URIBuilder(value2).setScheme(uri1.getScheme()).build();
+      return !uri1.relativize(uri2WithUri1Scheme).equals(uri2WithUri1Scheme);
+    } catch (final URISyntaxException e) {
+      return false;
+    }
   }
 }