+++ tiger-proxy/src/main/java/de/gematik/test/tiger/mockserver/mock/HttpState.java
@@ -91,7 +91,6 @@ public class HttpState {
   }
 
   public boolean handle(HttpRequest request) {
-
     request.withLogCorrelationId(UUIDService.getUUID());
     setPort(request);